+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0
 
 * Adds support for heatmap layers.
+* Adds options for gesture handling and tilt controls on web.
 
 ## 2.2.4
 
@@ -43,7 +44,7 @@
 
 ## 2.1.5
 
-Removes dependency on `meta`.
+* Removes dependency on `meta`.
 
 ## 2.1.4