+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 * Adds support for heatmap layers.
 * Adds options for gesture handling and tilt controls on web.
+## NEXT
+
+* Updates minimum Flutter version to 3.0.
+
+## 2.2.5
+
+* Updates code for stricter lint checks.
 
 ## 2.2.4